feat(run): fractional durations + run-phase announcements (#598)
Two Setup & Run improvements the loom composite runner was missing.

Fractional duration — a temporal composite can now run a float duration (e.g.
2.5 time units), not just whole ticks:
- broker `_drive` runs the whole ticks, then a final `composite.run(frac)` for
  the remainder (was `for step in range(int(steps))`, truncating the fraction).
- the request pipeline is float end-to-end: CompositeTestRunRequest.steps,
  composite_test_run_views (was `int(...)`), run_runner.RunRequest.steps.
- n_steps persists the float (sqlite); the progress bar clamps to 100%.
- the Duration field accepts decimals (step="any", parseFloat).

Run-phase announcements — the coarse status stays "running" through the whole
pipeline, so a new nullable `phase` sub-status tracks the stage:
- run_runner sets phase simulate → "rendering visualizations" → "analysis flush"
  (via new composite_runs.set_phase; migrated-in `phase` column).
- the status view returns `phase` while running (None once terminal).
- Setup & Run shows a pulsing "Analysis flush…" / "Rendering visualizations…"
  line for post-sim stages, and "Simulating — step N of M" during the sim.

Verified live: a steps=2.5 run is accepted + stored as n_steps=2.5 and writes
phase=simulating (the demo composite's own COPASI model is broken, unrelated).
7 new backend tests (drive fractions, phase roundtrip, terminal gate); loom 327
+ backend composite_runs suites green.

+++ vivarium_workbench/lib/composite_test_run_views.py
@@ -70,7 +70,9 @@ def composite_test_run(ws_root: Path, body: dict) -> tuple[dict, int]:
 
     spec_id = (body.get("id") or "").strip()
     overrides = body.get("overrides") or {}
-    steps = int(body.get("steps") or 5)
+    # float: a temporal composite may run a fractional duration (e.g. 2.5); an
+    # int here truncated it. A whole number stays whole (JSON 5 → 5.0 → runs 5).
+    steps = float(body.get("steps") or 5)
     label = (body.get("label") or "").strip() or auto_label(overrides)
     emit_paths = body.get("emit_paths") or []
     if not isinstance(emit_paths, list):
